refactor: ActivityTable: extract subcomponents as standalone components
